chore: remove _instance and use AlgoliaAnalytics instead
