+++ lib/node.ts
@@ -1,3 +1,9 @@
-import { getInstanceForNode } from "./_instance";
+import AlgoliaAnalytics from "./insights";
+import { getFunctionalInterface } from "./_getFunctionalInterface";
+import { getRequesterForNode } from "./utils/request";
 
-export default getInstanceForNode();
+const requestFn = getRequesterForNode();
+const instance = new AlgoliaAnalytics({ requestFn });
+const functionalInterface = getFunctionalInterface(instance);
+
+export default functionalInterface;
